# api/openapi.yaml
openapi: 3.0.3
info:
  title: Keycast HTTP API
  version: 1.0.0
  description: |
    # Keycast HTTP Signing API

    Fast HTTP signing for Nostr events without NIP-46 relay overhead. Keycast provides both traditional NIP-46 remote signing and a high-performance HTTP API for low-latency signing operations.

    ## Authentication

    Most endpoints require a UCAN token obtained from the `/auth/login` or `/auth/register` endpoints. Include the token in the `Authorization` header:

    ```
    Authorization: Bearer <your-jwt-token>
    ```

    ## Performance

    - **Fast Path**: ~10ms signing using cached handlers (unified mode)
    - **Slow Path**: ~50-100ms signing using database + KMS decryption

    ## Security Notes

    ⚠️ **Current Limitations**:
    - No permission validation on HTTP signing endpoint
    - No rate limiting (endpoints can be called unlimited times)
    - UCAN tokens valid for 24 hours

    These limitations are being addressed in upcoming releases.

  contact:
    name: Keycast Support
    url: https://github.com/rabble/keycast
  license:
    name: MIT
servers:
  - url: https://oauth.divine.video/api
    description: Production server
  - url: http://localhost:3000/api
    description: Local development
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: UCAN token from /auth/register or /auth/login
...
paths:
...
  /.well-known/nostr.json:
    get:
      summary: NIP-05 discovery
      description: |
        NIP-05 identifier verification endpoint.

        Query by username: `/.well-known/nostr.json?name=alice`
      tags:
        - Discovery
      parameters:
        - name: name
          in: query
          required: false
          description: Username to look up
          schema:
            type: string
            example: alice
      responses:
        '200':
          description: NIP-05 discovery response
          content:
            application/json:
              schema:
                type: object
                properties:
                  names:
                    type: object
                    additionalProperties:
                      type: string
                    example:
                      alice: "abc123..."
                  relays:
                    type: object
                    additionalProperties:
                      type: array
                      items:
                        type: string
                    example:
                      "abc123...": ["wss://relay.example.com"]
...